Rename: 컴포넌트 폴더를 아토믹 단위로 분리한다.
