+++ src/components/molecules/profile/Profile.tsx
@@ -2,6 +2,7 @@
 import { Image, ProfileLayout } from "./Profile.styled";
 import { renderProfile } from "./Profile.utils";
 
+
 type Profile = 'Default' | 'Portfolio' | 'Chat';
 
 type Props = {
