+++ src/components/organisms/editor/Editor.styled.tsx
