+++ src/components/organisms/modal/search-modal/SearchModal.styled.tsx
