+++ src/components/organisms/modal/search-modal/SearchModal.type.ts
