+++ src/components/organisms/preview/Preview.styled.tsx
